backend/requirements.txt
asgiref==3.3.1
boto3==1.17.7
botocore==1.20.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.4
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-richtextfield==1.6
django-storages==1.11.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
stripe==2.55.2
urllib3==1.26.3
whitenoise==5.2.0
